application/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
python-dotenv
pydantic[email]
requests
cryptography
passlib[argon2]
jinja2>=3.0.0
pyyaml>=6.0
pytest
pytest-asyncio
pytest-cov
python-multipart